scala scripts for icu4j

+++ comp-icu4j-vs-rbnfgo/batch_run_icu4j_cardinal.scala
@@ -18,14 +18,14 @@ val rbnf = new RuleBasedNumberFormat(new ULocale(args(0)), RuleBasedNumberFormat
 if (args.length>=2) {
   for (i <- 1 until args.length) {
     val s = args(i)
-    val n = s.toInt
+    val n = s.toLong
     val res = rbnf.format(n)
     Console.out.println(s + "\t" + res)
   }
 } else {
   var s = ""
   while ({s = StdIn.readLine(); s != null}) {
-    val n = s.toInt
+    val n = s.toLong
     val res = rbnf.format(n)
     Console.out.println(s + "\t" + res)
   }
