+++ comp-icu4j-vs-rbnfgo/batch_run_icu4j_ordinal.scala
@@ -22,14 +22,14 @@ val rbnf = new RuleBasedNumberFormat(new ULocale(args(0)), RuleBasedNumberFormat
 if (args.length>=3) {
   for (i <- 2 until args.length) {
     val s = args(i)
-    val n = s.toInt
+    val n = s.toLong
     val res = rbnf.format(n,ruleset)
     Console.out.println(s + "\t" + res)
   }
 } else {
   var s = ""
   while ({s = StdIn.readLine(); s != null}) {
-    val n = s.toInt
+    val n = s.toLong
     val res = rbnf.format(n,ruleset)
     Console.out.println(s + "\t" + res)
   }
